Refactoring io
